Resolve the three conflicts main's Chile ETL created against the Canada
adapters, all of them append-at-the-end collisions:

- highliner/core/density.py: Chile claimed bit 64, so ca_protected moves to
  32768. No Canada data exists yet, so no stored mask is invalidated.
- highliner/core/restrictions.py and
  frontend/src/lib/i18n/restrictionStrings.ts: keep both sides' layer entries.

+++ tests/highliner/server/repositories/test_density_store.py
@@ -29,7 +29,7 @@ def _write(path: Path, hist: list[tuple[int, int, int, int]]) -> None:
         off=np.array([0, len(hist)], dtype=np.int64),
         hl=np.array([row[0] for row in hist], dtype=np.int16),
         he=np.array([row[1] for row in hist], dtype=np.int16),
-        hm=np.array([row[2] for row in hist], dtype=np.int8),
+        hm=np.array([row[2] for row in hist], dtype=np.int16),
         hc=np.array([row[3] for row in hist], dtype=np.int32),
     )
